MasterFitness

+++ compon/src/main/java/fitness/sistem/compon/json_simple/SimpleRecordToJson.java
@@ -38,7 +38,7 @@ public class SimpleRecordToJson {
                     sb.append(quote + f.name + quoteColon + (Integer) f.value);
                     break;
                 case Field.TYPE_LONG :
-                    sb.append(quote + f.name + quoteColon + (Long) f.value);
+                    sb.append(quote + f.name + quoteColon + f.value);
                     break;
                 case Field.TYPE_DOUBLE :
                     sb.append(quote + f.name + quoteColon + (Double) f.value);
